Limiting number of hits to uint16_t due to possible screen overflow.

+++ firmware/application/apps/ble_rx_app.hpp
@@ -79,7 +79,7 @@ struct BleRecentEntry {
     std::string timestamp;
     std::string dataString;
     std::string nameString;
-    uint64_t numHits;
+    uint16_t numHits;
 
     BleRecentEntry()
         : BleRecentEntry{0} {
